requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.2
fastapi-jwt-auth==0.5.0
fastapi-sessions==0.3.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==1.7.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
wheel==0.44.0
mega.py
python-dotenv
